refactor: 도메인별 state 분리
